fix(#11): Ordered all imports and moved all at the top of the files. Also deleted tmp.txt because it was being unused
