+++ backend/plugins/base.py
@@ -4,13 +4,13 @@ Base plugin interface for ingestion strategies.
 This module defines the base plugin interface and registry for document ingestion strategies.
 """
 
+# Python Libraries
 import abc
 import importlib
-import inspect
 import os
 import pkgutil
-from enum import Enum, auto
-from typing import Dict, List, Type, Any, Optional, Set, Tuple, Union
+from enum import Enum
+from typing import Any, Dict, List, Optional, Set, Type, Union
 
 
 class ChunkUnit(str, Enum):
