+++ backend/schemas/system.py
@@ -1,5 +1,8 @@
+# Python Libraries
+from typing import Any, Dict
+
+# Third-Party Libraries
 from pydantic import BaseModel, Field
-from typing import Dict, Any
 
 # Response models
 class HealthResponse(BaseModel):
